+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/JsContributionProvidersBoxTestGenerated.java
@@ -2791,6 +2791,12 @@ public class JsContributionProvidersBoxTestGenerated extends AbstractJsContribut
     }
 
     @Test
+    @TestMetadata("InlinedProvidesAcrossModulesBytecodeCheck.kt")
+    public void testInlinedProvidesAcrossModulesBytecodeCheck() {
+      run("InlinedProvidesAcrossModulesBytecodeCheck.kt");
+    }
+
+    @Test
     @TestMetadata("InlinedProvidesBytecodeCheck.kt")
     public void testInlinedProvidesBytecodeCheck() {
       run("InlinedProvidesBytecodeCheck.kt");